fastapi_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai==0.28.1
python-dotenv==1.0.0
pydantic==2.4.2
